pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aethergrid"
version = "2.0.0"
description = "Decentralized OS & Task Orchestration Platform with Raft consensus, fencing tokens, and log-based leases"
readme = "ARCHITECTURE.MD"
license = {text = "MIT"}
authors = [
    {name = "AetherGrid Contributors"},
]
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.15.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Distributed Computing",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0",
    "ruff>=0.0.260",
]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
aether = "aethergrid.cli:main"

[project.urls]
Homepage = "https://github.com/aethergrid/aethergrid"
Documentation = "https://github.com/aethergrid/aethergrid/blob/main/ARCHITECTURE.MD"
Repository = "https://github.com/aethergrid/aethergrid"

[tool.hatch.build.targets.wheel]
packages = ["aethergrid"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
